Handle key backup recover banner preferences
